02-agentcore-memory/deployment/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents
strands-agents-tools
aws-opentelemetry-distro
